requirements.txt
django-allauth
djangorestframework
markdown
django-filter
kafka-python